+++ ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/DefaultShieldsSectionView.swift
@@ -15,20 +15,7 @@ import Strings
 import SwiftUI
 
 struct DefaultShieldsSectionView: View {
-  enum CookieAlertType: String, Identifiable {
-    case confirm
-    case failed
-
-    var id: String { rawValue }
-  }
-
   @ObservedObject var settings: AdvancedShieldsSettings
-  @State private var cookieAlertType: CookieAlertType?
-  /// If we should force show the Block All Cookies row.
-  /// If a user disables the toggle with the feature flag disabled, we don't
-  /// want the row to disappear on the user. So if we are showing the row
-  /// when the view opens, it should remain visible until the view is dismissed.
-  @State private var showBlockAllCookies = false
 
   var body: some View {
     Section {
@@ -91,51 +78,6 @@ struct DefaultShieldsSectionView: View {
         option: Preferences.Shields.blockScripts
       )
 
-      if showBlockAllCookies || FeatureList.kBlockAllCookiesToggle.enabled
-        || Preferences.Privacy.blockAllCookies.value
-      {
-        OptionToggleView(
-          title: Strings.blockAllCookies,
-          subtitle: Strings.blockCookiesDescription,
-          option: Preferences.Privacy.blockAllCookies,
-          onChange: { newValue in
-            if newValue {
-              cookieAlertType = .confirm
-            } else {
-              Task {
-                await toggleCookieSetting(with: false)
-              }
-            }
-          }
-        )
-        .alert(item: $cookieAlertType) { cookieAlertType in
-          switch cookieAlertType {
-          case .confirm:
-            return Alert(
-              title: Text(Strings.blockAllCookiesAction),
-              message: Text(Strings.blockAllCookiesAlertInfo),
-              primaryButton: .default(
-                Text(Strings.blockAllCookiesAction),
-                action: {
-                  Task {
-                    await toggleCookieSetting(with: true)
-                  }
-                }
-              ),
-              secondaryButton: .cancel(
-                Text(Strings.cancelButtonTitle),
-                action: {
-                  Preferences.Privacy.blockAllCookies.value = false
-                }
-              )
-            )
-          case .failed:
-            return Alert(
-              title: Text(Strings.blockAllCookiesFailedAlertMsg)
-            )
-          }
-        }
-      }
       OptionToggleView(
         title: Strings.fingerprintingProtection,
         subtitle: Strings.fingerprintingProtectionDescription,
@@ -183,33 +125,6 @@ struct DefaultShieldsSectionView: View {
       Text(Strings.shieldsDefaultsFooter)
     }
     .listRowBackground(Color(.secondaryBraveGroupedBackground))
-    .onAppear {
-      showBlockAllCookies = Preferences.Privacy.blockAllCookies.value
-    }
-  }
-
-  private func toggleCookieSetting(with status: Bool) async {
-    do {
-      try await AsyncFileManager.default.setWebDataAccess(atPath: .cookie, lock: status)
-      try await AsyncFileManager.default.setWebDataAccess(atPath: .websiteData, lock: status)
-
-      if Preferences.Privacy.blockAllCookies.value != status {
-        Preferences.Privacy.blockAllCookies.value = status
-      }
-    } catch {
-      Logger.module.error("Failed to change web data access to \(status)")
-      if status {
-        // Revert the changes. Not handling success here to avoid a loop.
-        try? await AsyncFileManager.default.setWebDataAccess(atPath: .cookie, lock: false)
-        try? await AsyncFileManager.default.setWebDataAccess(atPath: .websiteData, lock: false)
-
-        if Preferences.Privacy.blockAllCookies.value != false {
-          Preferences.Privacy.blockAllCookies.value = false
-        }
-
-        cookieAlertType = .failed
-      }
-    }
   }
 }
